csv_analyzer/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.0.0
pandas>=1.3.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
openpyxl>=3.0.0
requests>=2.25.0
sqlite3